+++ Den/Views/Sidebar/SidebarToolbar.swift
@@ -17,7 +17,7 @@ struct SidebarToolbar: CustomizableToolbarContent {
 
     @Binding var currentProfileID: String?
     @Binding var detailPanel: DetailPanel?
-    @Binding var feedRefreshTimeout: Double
+    @Binding var feedRefreshTimeout: Int
     @Binding var isEditing: Bool
     @Binding var refreshing: Bool
     @Binding var showingExporter: Bool
